Keep knowledge command path logic easier to extend
Extract the path normalization, source collection, and result-shaping
helpers into a dedicated module so the knowledge command stays focused
on orchestration rather than low-level ingest bookkeeping.

Constraint: Preserve CLI knowledge-ingest behavior and existing command tests
Constraint: Repo pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Splitting parser execution helpers in the same pass | broader than one queue item needs
Confidence: high
Scope-risk: narrow
Directive: Keep command-helpers.ts limited to pure/path/source helpers; leave parser execution and command orchestration in command.ts
Tested: deno test --no-check --allow-all cli/commands/knowledge/command.test.ts; deno fmt --check cli/commands/knowledge/command.ts cli/commands/knowledge/command-helpers.ts; deno lint cli/commands/knowledge/command.ts cli/commands/knowledge/command-helpers.ts; git diff --check
Not-tested: Pre-commit verify:quick (fails on existing cli boundary violations in cli/commands/extension/*.ts)
