+++ cli/commands/knowledge/command.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 type SafeParseResult<T> = { success: true; data: T } | { success: false; error: z.ZodError };
 import { type CommandResult, createFileSystem, getEnv, runCommand } from "veryfront/platform";
-import { basename, extname, join, normalize, relative } from "veryfront/platform/path";
+import { basename } from "veryfront/platform/path";
 import { withSpan } from "veryfront/observability/otlp-setup";
 import { cliLogger } from "#cli/utils";
 import { type ApiClient, createApiClient, resolveConfigWithAuth } from "#cli/shared/config";
@@ -10,9 +10,10 @@ import { getStringArg } from "../../shared/parsed-args.ts";
 import { downloadUploadToFile, listAllUploads, type UploadItem } from "../uploads/command.ts";
 import { putRemoteFileFromLocal } from "../files/command.ts";
 import { knowledgeIngestPythonSource } from "./parser-source.ts";
+import * as commandHelpers from "./command-helpers.ts";
 import { createJobUserLogger, type Logger, serverLogger } from "veryfront/utils";
 import { writeJobResultIfConfigured } from "../../utils/write-job-result.ts";
-import { classifyKnowledgeDirectoryPath, classifyKnowledgeSourcePath } from "./source-policy.ts";
+import { classifyKnowledgeSourcePath } from "./source-policy.ts";
 import {
   buildKnowledgeIngestJobResult,
   type KnowledgeIngestFailedFileResult,
@@ -171,106 +172,8 @@ export function parseKnowledgeIngestArgs(
   });
 }
 
-export function normalizeKnowledgeInputPath(inputPath: string): string {
-  const normalizedPath = normalize(inputPath).replace(/^\/+/, "").replace(/\\/g, "/");
-  if (!normalizedPath || normalizedPath.startsWith("..") || normalizedPath.startsWith("/")) {
-    throw new Error(`Invalid knowledge input path: ${inputPath}`);
-  }
-  return normalizedPath;
-}
-
-export function normalizeProjectUploadPath(inputPath: string): string {
-  return normalizeKnowledgeInputPath(inputPath);
-}
-
-export function formatKnowledgeUploadSource(uploadPath: string): string {
-  const normalizedPath = normalizeKnowledgeInputPath(uploadPath);
-  return normalizedPath === "uploads" || normalizedPath.startsWith("uploads/")
-    ? normalizedPath
-    : `uploads/${normalizedPath}`;
-}
-
-function resolveExplicitUploadPath(inputPath: string): string {
-  const normalizedInput = normalizeKnowledgeInputPath(inputPath);
-  const displayInput = inputPath.replace(/\\/g, "/");
-  const uploadPath = normalizeProjectUploadPath(inputPath);
-  if (
-    !uploadPath || uploadPath === "uploads" || normalizedInput === "uploads" ||
-    normalizedInput.endsWith("/")
-  ) {
-    throw new Error(
-      `Directory upload references require --path <prefix> --all: ${displayInput}`,
-    );
-  }
-  return uploadPath;
-}
-
-export function isLikelyLocalPath(value: string): boolean {
-  return value.startsWith("/") || value.startsWith("./") || value.startsWith("../") ||
-    /^[A-Za-z]:[\\/]/.test(value);
-}
-
-function isProjectUploadReference(value: string): boolean {
-  if (isLikelyLocalPath(value)) return false;
-  const normalizedValue = normalize(value).replace(/\\/g, "/").replace(/^\/+/, "");
-  return normalizedValue === "uploads" || normalizedValue.startsWith("uploads/");
-}
-
-function slugify(value: string): string {
-  return value.toLowerCase().replace(/[^a-z0-9]+/g, "-").replace(/^-+|-+$/g, "") || "document";
-}
-
-/**
- * Strip the chat-upload prefix that Studio prepends to uploaded files.
- *
- * Studio stores uploads with a generated prefix like:
- *   chat-<uuid>-<timestamp>-<shortid>-<original-filename>
- *
- * This function removes the prefix so knowledge files use the clean
- * original filename (e.g. "agents" instead of
- * "chat-909d3dbc-5a9a-4156-97e4-bcceb5c2ec0d-1773942180291-fv1qg5-agents").
- */
-const CHAT_UPLOAD_PREFIX_RE =
-  /^chat-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}-\d+-[a-z0-9]+-/i;
-
-export function stripChatUploadPrefix(name: string): string {
-  return name.replace(CHAT_UPLOAD_PREFIX_RE, "");
-}
-
 function defaultOutputRoot(): Promise<string> {
   return Deno.makeTempDir({ prefix: "veryfront-knowledge-" });
-}
-
-export function resolveKnowledgeDownloadOutputDir(outputDir: string): string {
-  return join(outputDir, ".uploads");
-}
-
-function createSkippedKnowledgeSource(input: {
-  source: string;
-  localSourcePath?: string | null;
-  message: string;
-  reason: KnowledgeIngestSkippedFileResult["reason"];
-}): KnowledgeIngestSkippedFileResult {
-  return {
-    source: input.source,
-    localSourcePath: input.localSourcePath ?? null,
-    message: input.message,
-    reason: input.reason,
-  };
-}
-
-function createFailedKnowledgeSource(input: {
-  source: string;
-  localSourcePath: string;
-  message: string;
-  reason: KnowledgeIngestFailedFileResult["reason"];
-}): KnowledgeIngestFailedFileResult {
-  return {
-    source: input.source,
-    localSourcePath: input.localSourcePath,
-    message: input.message,
-    reason: input.reason,
-  };
 }
 
 function classifySourceOrSkip(input: {
@@ -282,180 +185,51 @@ function classifySourceOrSkip(input: {
     return null;
   }
 
-  return createSkippedKnowledgeSource({
+  return {
     source: input.source,
-    localSourcePath: input.localSourcePath,
+    localSourcePath: input.localSourcePath ?? null,
     message: decision.message,
     reason: decision.reason,
-  });
+  };
 }
 
-function classifyDirectoryOrSkip(input: {
-  source: string;
-}): KnowledgeIngestSkippedFileResult | null {
-  const decision = classifyKnowledgeDirectoryPath(input.source);
-  if (decision.kind === "ingest") {
-    return null;
-  }
+export const normalizeKnowledgeInputPath = commandHelpers.normalizeKnowledgeInputPath;
+export const normalizeProjectUploadPath = commandHelpers.normalizeProjectUploadPath;
+export const formatKnowledgeUploadSource = commandHelpers.formatKnowledgeUploadSource;
+export const isLikelyLocalPath = commandHelpers.isLikelyLocalPath;
+export const stripChatUploadPrefix = commandHelpers.stripChatUploadPrefix;
+export const resolveKnowledgeDownloadOutputDir = commandHelpers.resolveKnowledgeDownloadOutputDir;
+export const buildSuggestedSlug = commandHelpers.buildSuggestedSlug;
+export const ensureUniqueSlugs = commandHelpers.ensureUniqueSlugs;
+export const deriveKnowledgeRemotePath = commandHelpers.deriveKnowledgeRemotePath;
+export const createKnowledgeIngestResult = commandHelpers.createKnowledgeIngestResult;
 
-  return createSkippedKnowledgeSource({
-    source: input.source,
-    localSourcePath: null,
-    message: decision.message,
-    reason: decision.reason,
-  });
+function resolveExplicitUploadPath(inputPath: string): string {
+  return commandHelpers.resolveExplicitUploadPath(inputPath);
+}
+
+function isProjectUploadReference(value: string): boolean {
+  return commandHelpers.isProjectUploadReference(value);
+}
+
+function createFailedKnowledgeSource(input: {
+  source: string;
+  localSourcePath: string;
+  message: string;
+  reason: KnowledgeIngestFailedFileResult["reason"];
+}): KnowledgeIngestFailedFileResult {
+  return commandHelpers.createFailedKnowledgeSource(input);
 }
 
 async function collectLocalFiles(
   root: string,
   recursive: boolean,
 ): Promise<KnowledgeSourceCollection> {
-  const fs = createFileSystem();
-  const stat = await fs.stat(root);
-  if (stat.isFile) {
-    const skipped = classifySourceOrSkip({ source: root, localSourcePath: root });
-    return skipped == null
-      ? {
-        sources: [{ kind: "local", input: root, localPath: root }],
-        skipped: [],
-      }
-      : {
-        sources: [],
-        skipped: [skipped],
-      };
-  }
-  if (!stat.isDirectory) {
-    return { sources: [], skipped: [] };
-  }
-
-  const skippedRootDirectory = classifyDirectoryOrSkip({ source: root });
-  if (skippedRootDirectory != null) {
-    return {
-      sources: [],
-      skipped: [skippedRootDirectory],
-    };
-  }
-
-  const collection: KnowledgeSourceCollection = {
-    sources: [],
-    skipped: [],
-  };
-  async function walk(dir: string): Promise<void> {
-    for await (const entry of fs.readDir(dir)) {
-      const entryPath = join(dir, entry.name);
-      if (entry.isDirectory) {
-        const skipped = classifyDirectoryOrSkip({ source: entryPath });
-        if (skipped != null) {
-          collection.skipped.push(skipped);
-          continue;
-        }
-        if (recursive) await walk(entryPath);
-        continue;
-      }
-
-      if (!entry.isFile) {
-        continue;
-      }
-
-      const skipped = classifySourceOrSkip({ source: entryPath, localSourcePath: entryPath });
-      if (skipped != null) {
-        collection.skipped.push(skipped);
-        continue;
-      }
-
-      collection.sources.push({ kind: "local", input: root, localPath: entryPath });
-    }
-  }
-
-  await walk(root);
-  collection.sources.sort((left, right) => left.localPath.localeCompare(right.localPath));
-  collection.skipped.sort((left, right) => left.source.localeCompare(right.source));
-  return collection;
+  return commandHelpers.collectLocalFiles(root, recursive);
 }
 
 function buildSourceReference(source: KnowledgeSource): string {
-  return source.kind === "upload"
-    ? formatKnowledgeUploadSource(source.uploadPath)
-    : source.localPath;
-}
-
-export function buildSuggestedSlug(source: KnowledgeSource, index: number): string {
-  const normalized = normalize(
-    source.kind === "upload" ? source.uploadPath : source.localPath,
-  ).replace(/\\/g, "/");
-
-  let stripped: string;
-  if (source.kind === "upload") {
-    stripped = normalized
-      .replace(/^\/workspace\/uploads\//, "")
-      .replace(/^\/workspace\//, "")
-      .replace(/^uploads\//, "")
-      .replace(/\.[^.]+$/, "");
-  } else if (normalized.startsWith("/workspace/uploads/")) {
-    stripped = normalized.replace(/^\/workspace\/uploads\//, "").replace(/\.[^.]+$/, "");
-  } else if (normalized.startsWith("/workspace/")) {
-    stripped = normalized.replace(/^\/workspace\//, "").replace(/\.[^.]+$/, "");
-  } else if (normalized.startsWith("/")) {
-    stripped = basename(normalized, extname(normalized));
-  } else {
-    stripped = normalized.replace(/\.[^.]+$/, "");
-  }
-
-  // Strip Studio's chat-upload prefix from the filename portion so that
-  // knowledge files use the original clean filename.
-  const lastSlash = stripped.lastIndexOf("/");
-  if (lastSlash >= 0) {
-    const dir = stripped.slice(0, lastSlash + 1);
-    const file = stripChatUploadPrefix(stripped.slice(lastSlash + 1));
-    stripped = file ? `${dir}${file}` : stripped;
-  } else {
-    stripped = stripChatUploadPrefix(stripped) || stripped;
-  }
-
-  return slugify(stripped || basename(normalized, extname(normalized)) || `document-${index + 1}`);
-}
-export function ensureUniqueSlugs(sources: KnowledgeSource[]): string[] {
-  const counts = new Map<string, number>();
-  return sources.map((source, index) => {
-    const baseSlug = buildSuggestedSlug(source, index);
-    const nextCount = (counts.get(baseSlug) ?? 0) + 1;
-    counts.set(baseSlug, nextCount);
-    return nextCount === 1 ? baseSlug : `${baseSlug}-${nextCount}`;
-  });
-}
-
-export function deriveKnowledgeRemotePath(
-  outputPath: string,
-  outputDir: string,
-  knowledgePath: string,
-): string {
-  const relativeOutputPath = relative(outputDir, outputPath).replace(/\\/g, "/");
-  if (!relativeOutputPath || relativeOutputPath.startsWith("..")) {
-    throw new Error(`Output path is outside output directory: ${outputPath}`);
-  }
-  const prefix = normalizeKnowledgeInputPath(knowledgePath);
-  const normalizedRelative = normalize(relativeOutputPath).replace(/^\/+/, "");
-  return `${prefix}/${normalizedRelative}`.replace(/\\/g, "/");
-}
-
-export function createKnowledgeIngestResult(input: {
-  source: string;
-  localSourcePath: string;
-  outputPath: string;
-  remotePath: string;
-  parser: Pick<KnowledgeParserResult, "slug" | "stats" | "warnings" | "source_type" | "summary">;
-}): KnowledgeIngestFileResult {
-  return {
-    source: input.source,
-    localSourcePath: input.localSourcePath,
-    outputPath: input.outputPath,
-    remotePath: input.remotePath,
-    slug: input.parser.slug,
-    sourceType: input.parser.source_type,
-    summary: input.parser.summary,
-    stats: input.parser.stats,
-    warnings: input.parser.warnings,
-  };
+  return commandHelpers.buildSourceReference(source);
 }
 
 export async function runKnowledgeParser(input: {
